dernieres routes + numeroSS 13 nombres obligatoire

+++ AppHopital/src/Controller/PostVersApiController.php
@@ -17,14 +17,24 @@ class PostVersApiController extends AbstractController
     #[Route('/post/patient', name: 'post_patient')]
     public function PostPatient():Response
     {
+        $session = $request->getSession();
+        $role = $session->get('role');
+        if ($role == 'ROLE_ADMIN' || $role == 'ROLE_USER'){
+            $infirmiers = new TableauInfirmier;
+            $infirmiers->DeleteInfirmier(47);
 
-        $infirmiers = new TableauInfirmier;
-        $infirmiers->DeleteInfirmier(47);
+            return $this->render('post_vers_api/index.html.twig', [
+                'controller_name' => 'PostVersApiControllerController',
+                'title'=> 'Accueil',
+            ]);
+        }else{
+            return $this->render('access_denied/index.html.twig', [
+                'controller_name' => 'PostVersApiControllerController',
+                'error' => "Vous n'êtes pas autorisé à aller sur cette page"
+            ]);
+        }
 
-        return $this->render('post_vers_api/index.html.twig', [
-            'controller_name' => 'PostVersApiControllerController',
-            'title'=> 'Accueil',
-        ]);
+
 
     }
 }
